Player Controller Refactor
Refactored so each piece of the player controller is in its on script.  Will keep things cleaner when implimenting all the different elemental movement things.
