+++ Scripts/Character/attack.cs
@@ -1,0 +1,7 @@
+using Godot;
+using System;
+
+public partial class attack : Node
+{
+    
+}
