+++ game.tscn
@@ -1,7 +1,7 @@
 [gd_scene load_steps=6 format=3 uid="uid://30pn8csybo1i"]
 
 [ext_resource type="Texture2D" uid="uid://bbluoyi6e8jva" path="res://Assets/Art/Level/background.png" id="1_e051u"]
-[ext_resource type="Script" path="res://Player.cs" id="2_q20p0"]
+[ext_resource type="Script" path="res://Scripts/Character/Player.cs" id="2_q20p0"]
 [ext_resource type="Texture2D" uid="uid://bk5egf1oa654v" path="res://Assets/Art/Player/temp.png" id="3_y7226"]
 
 [sub_resource type="RectangleShape2D" id="RectangleShape2D_7fvru"]
@@ -15,7 +15,7 @@ size = Vector2(74, 197)
 [node name="Level" type="StaticBody2D" parent="."]
 
 [node name="Sprite2D" type="Sprite2D" parent="Level"]
-position = Vector2(574.813, 326.25)
+position = Vector2(576.813, 325.25)
 scale = Vector2(0.643066, 0.638184)
 texture = ExtResource("1_e051u")
